requirements.txt
torch
torchvision
Pillow
black
flake8
# aiohttp
aiohttp
aiohttp_jinja2
# flask
flask
